+++ packages/apps/campaign/src/apps/create-campaign/screens/Scenarios.tsx
@@ -12,8 +12,8 @@ const ScenarioItem = (props: { scenario: Scenario; onPress: (name: Scenario) => 
 				return Styles["red-bullet"];
 			case "road-to-paris":
 				return Styles["road-to-paris"];
-			case "desert-thunder":
-				return Styles["desert-thunder"];
+			case "operation-northern-shield":
+				return Styles["operation-northern-shield"];
 			default:
 				return;
 		}